[PL-2055]: Unique account name (#8144)
Add function to suggest unique account names

Unique account names are created by appending 4
random digits to the given account name. The
check for duplicated account name is case insensitive.
The random generator also has a limit of 20, if it cannot
generate a unique name is fails with an error.